Remove [non-]illumos duplication for Port[Manager]
