+++ illumos-utils/src/opte/non_illumos.rs
@@ -6,13 +6,6 @@
 
 use slog::Logger;
 
-mod port;
-mod port_manager;
-
-pub use port::Port;
-pub use port_manager::PortManager;
-pub use port_manager::PortTicket;
-
 use crate::addrobj::AddrObject;
 
 #[derive(thiserror::Error, Debug)]
